Replace C1 controls with C0 to emit OSC 777 from PS0
C1 controls are supported by fewer terminals, and using them here
causes a problem in Alacritty (and presumably quite a few others) if
it's launched from a VTE-based terminal. See:
https://github.com/jwilm/alacritty/issues/3105.

Support for C1 controls might also get removed from VTE itself. See:
https://gitlab.gnome.org/GNOME/vte/issues/209.

Note that there was a typo in the commit message of the patch that
introduced support for preeexec notifications from an interactive
shell. While it was correct in saying that it's using C1 controls,
because C0 was causing spurious characters to show up, the C0 example
was wrongly using C1 controls.

https://bugzilla.redhat.com/show_bug.cgi?id=1783802